added all python demos
