--- python_demos/data/defect_detection_model.dwm
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:0af75584f9e7f352025ea73f41900d1f19141a2d8aa000c21763c8dbbf804553
-size 299574078
