+++ python_demos/data/supervised_defect_segmentation_model.dwm
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:4c6b451eae98bb852935bd47791922b639814577903218bc0866b6a6e5d2a8d9
+size 322316254
