+++ python_demos/data/unsupervised_defect_segmentation_model.dwm
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:56beace5a4450347b6dc17d0032cdf30083429637093f18e4375c30c4109c035
+size 96535437
