+++ python_demos/requirements.txt
@@ -1,1 +1,2 @@
 opencv-python
+PyQt5>=5.15.0
